gemini-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.1
google-genai==1.47.0
numpy==1.24.3
supabase==2.0.2